requirements.txt
pandas
scikit-learn
pyyaml
nltk
datasets
transformers
torch
"transformers[torch]"